feat(search): Add support for Omnibar
This is a requirement from the original extension that I missed during
the rewrite. Now users can type search queries into the Chrome omnibar,
and it can open up the search interface or deep link directly to ATT&CK
website.

As part of this, I needed to move manifest.json from the top level
directory into public/ so that the URL path is the same when loading
ATT&CK data from the worker script as it is from the popup script.