Increase test coverage
This requires some refactoring to work around globals like
chrome.storage and the fetch() API, so I didn't get the coverage as high
as I wanted to, but at least got over the hump of mocking out the
storage API and fetch API. This commit includes proof-of-concept tests
for one component (App) and several library features (e.g. bookmarks and
formats).