Show "open in new tab" in contex menu
Previously this item was only shown when text is selected; now it is
shown even when no text is selected.

Also cleaned up the "Go to selected ATT&CK object" menu item--refactored
the code a bit and put the object ID into the menu text so that it's
more transparent about what the menu item will do when you click on it.