Remove "activeTab" permission from manifest
The most recent submission in the chrome web store was rejected due to
this permission not being used. I didn't realize that it is intended to
replace host permission, not complement it. I think we need host permission
in order to make the context menu work correctly--i.e. enable/disable the
"goto selected technique" item--so I'm keeping that one and getting rid
of activeTab.