Fix the implementation of ATT&CK domains
The bookmarks->navigator export feature wasn't really originally
designed correctly because it let the user type any arbitrary domain and
didn't check that the exported objects belonged in that domain. This
commit forces a valid domain and filters out techniques that are not
part of that domain when exporting to navigator.