Improve UX for search filters (#39)
As suggested in the issue thread, add a vertical line to clarify that
the object type filters are separate from the domain/deprecated filters.
Also add a bit of text under each to explain.

I also noticed that the earlier fix in storage.js was not working in the
chrome extension code path, so I fixed that as well.